RAG with Groq Inference Engine/requirements.txt
langchain
streamlit
ollama
langchain_community
python-dotenv
langserve
fastapi
uvicorn
ipykernel
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
wikipedia
arxiv
langchainhub
groq
langchain-groq
cassio 
beautifulsoup4